Revert "Show listen transcript immediately; gate audio behind a play button"
This reverts commit 067786a9d963205eede5c54d536fc885c8efb475.
