+++ src/ListenViewer.tsx
@@ -1,15 +1,14 @@
-// ListenViewer: a self-contained pane that lets a viewer follow a live, AI
-// speech-to-speech translation for a single language, produced by a Gemini Live
-// translator bot in the session's LiveKit room. Kept deliberately isolated (lazy
-// LiveKit import, own error/empty states) so any failure here is contained to
-// this pane and never disturbs the outline / slide views.
+// ListenViewer: a self-contained pane that lets a viewer hear live
+// speech-to-speech translated audio for a single language, produced by a Gemini
+// Live translator bot in the session's LiveKit room. Kept deliberately isolated
+// (lazy LiveKit import, own error/empty states) so any failure here is contained
+// to this pane and never disturbs the outline / slide views.
 //
-// The transcript shows immediately and unconditionally (it reads the shared Yjs
-// doc, no LiveKit needed). To keep that transcript flowing this pane joins the
-// LiveKit room on mount — that both spins up the translator bot and keeps the
-// session healthy — but the translated *audio* stays muted until the listener
-// presses the play button. In other words: read for free, opt in to hear.
-import { useEffect, useState } from "react";
+// The transcript is read from the shared Yjs doc (`liveTranscript-{code}`), which
+// the server-side translator bridge writes (one delta at a time, as a continuous
+// stream). That gives late joiners the full history and is the single source of
+// truth — there is no separate ephemeral interim line.
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import {
   LiveKitRoom,
   RoomAudioRenderer,
@@ -19,8 +18,8 @@ import {
 import "@livekit/components-styles";
 import { Track } from "livekit-client";
 import { LANGUAGE_BCP47, useStrings } from "./useLocale";
+import { useAsPlainText } from "./yjsUtils";
 import { LISTEN_ORIGINAL_CODE, DEFAULT_LISTEN_CODE } from "./listenLanguages";
-import { LiveTranscript } from "./LiveTranscript";
 import { getDocId } from "./getDocId";
 
 const ORGANIZER_PREFIX = "organizer-";
@@ -44,51 +43,79 @@ interface TokenResp {
   error?: string;
 }
 
-// Runs inside <LiveKitRoom>: report whether the speaker is present, and — only
-// while the listener has audio enabled — subscribe to and play the chosen audio
-// (the translator bot, or the speaker's original mic). The transcript itself is
-// rendered by the parent, outside the room.
-function ListenAudio({
+// One finalized transcript paragraph. Mounting fresh (only appended segments do)
+// plays a one-shot highlight animation, so new text is gently emphasized without
+// any diffing — the Yjs text is append-only.
+function TranscriptSegment({ text, isNew }: { text: string; isNew: boolean }) {
+  return (
+    <p className={`my-2 ${isNew ? "transcript-new" : ""}`}>{text}</p>
+  );
+}
+
+// Runs inside <LiveKitRoom>: subscribe only to the chosen audio (translator bot,
+// or the speaker's original mic), render the stable transcript, and report whether
+// the speaker is present.
+function ListenInner({
   docId,
+  langCode,
   releaseTarget,
   translatorIdentity,
   isOriginal,
-  audioOn,
-  onToggleAudio,
 }: {
   docId: string;
+  langCode: string;
   releaseTarget: string;
   translatorIdentity: string | null;
   isOriginal: boolean;
-  audioOn: boolean;
-  onToggleAudio: () => void;
 }) {
   const s = useStrings();
   const room = useRoomContext();
   const remoteParticipants = useRemoteParticipants();
+  const [finalized] = useAsPlainText(`liveTranscript-${langCode}`);
+  const scrollRef = useRef<HTMLDivElement | null>(null);
+  const atBottomRef = useRef(true);
 
   const speakerPresent = remoteParticipants.some((p) =>
     p.identity.startsWith(ORGANIZER_PREFIX)
   );
 
-  // Subscribe to the audio we want only while audio is enabled: the translator
-  // bot for a translation, or the speaker's raw mic for "Original / English".
-  // autoSubscribe is off, so we drive this explicitly. Re-running on participant
-  // changes is essential for late joiners — the track is already published, so no
-  // per-track event fires for us.
+  // Subscribe to the audio we want: the translator bot for a translation, or the
+  // speaker's raw mic for "Original / English". autoSubscribe is off, so we drive
+  // this explicitly. Re-running on participant changes is essential for late
+  // joiners — the track is already published, so no per-track event fires for us.
   useEffect(() => {
     if (!room) return;
     for (const participant of remoteParticipants) {
-      const wanted =
-        audioOn &&
-        (isOriginal
-          ? participant.identity.startsWith(ORGANIZER_PREFIX)
-          : participant.identity === translatorIdentity);
+      const wanted = isOriginal
+        ? participant.identity.startsWith(ORGANIZER_PREFIX)
+        : participant.identity === translatorIdentity;
       for (const [, pub] of participant.trackPublications) {
         if (pub.kind === Track.Kind.Audio) pub.setSubscribed(wanted);
       }
     }
-  }, [room, translatorIdentity, isOriginal, remoteParticipants, audioOn]);
+  }, [room, translatorIdentity, isOriginal, remoteParticipants]);
+
+  const segments = useMemo(
+    () => finalized.split("\n\n").map((t) => t.trim()).filter(Boolean),
+    [finalized]
+  );
+
+  // Segments present at first render aren't animated; only later-appended ones are.
+  // Captured once via a lazy initializer (reading a ref during render is unsafe).
+  const [baselineCount] = useState(() => segments.length);
+
+  // Auto-scroll only when the reader is already near the bottom, so reading older
+  // text isn't yanked away when new text arrives. Runs after every render (cheap);
+  // it only moves the scroll position when already pinned to the bottom.
+  const onScroll = () => {
+    const el = scrollRef.current;
+    if (!el) return;
+    atBottomRef.current = el.scrollHeight - el.scrollTop - el.clientHeight < 80;
+  };
+  useEffect(() => {
+    const el = scrollRef.current;
+    if (el && atBottomRef.current) el.scrollTop = el.scrollHeight;
+  });
 
   // Decrement the listener count when this pane goes away (unmount or tab close)
   // so idle translator bots tear down. (Best-effort; the server's presence reaper
@@ -108,6 +135,8 @@ function ListenAudio({
     };
   }, [docId, releaseTarget]);
 
+  const hasContent = segments.length > 0;
+
   return (
     <>
       <RoomAudioRenderer />
@@ -120,17 +149,25 @@ function ListenAudio({
         <span className="text-gray-600 dark:text-gray-300">
           {speakerPresent ? s.liveListening : s.waitingForSpeaker}
         </span>
-        <div className="flex-1" />
-        <button
-          type="button"
-          aria-pressed={audioOn}
-          className={`px-3 py-1 rounded-full text-white shadow ${
-            audioOn ? "bg-gray-500 hover:bg-gray-600" : "bg-blue-500 hover:bg-blue-600"
-          }`}
-          onClick={onToggleAudio}
-        >
-          {audioOn ? `🔇 ${s.stopAudio}` : `🔊 ${s.listenLive}`}
-        </button>
+      </div>
+      <div
+        ref={scrollRef}
+        onScroll={onScroll}
+        className="flex-1 overflow-auto text-gray-800 dark:text-gray-100"
+      >
+        {hasContent ? (
+          <div className="mx-auto max-w-[60ch] text-lg leading-relaxed">
+            {segments.map((seg, i) => (
+              <TranscriptSegment
+                key={`${i}-${seg.slice(0, 16)}`}
+                text={seg}
+                isNew={i >= baselineCount}
+              />
+            ))}
+          </div>
+        ) : (
+          <span className="italic text-gray-400">{s.waitingForSpeech}</span>
+        )}
       </div>
     </>
   );
@@ -147,109 +184,95 @@ export function ListenViewer({ language }: { language: string }) {
   const docId = getDocId();
   const [conn, setConn] = useState<ConnectInfo | null>(null);
   const [error, setError] = useState<string | null>(null);
-  // Audio is opt-in: the transcript flows without it, the button toggles sound.
-  const [audioOn, setAudioOn] = useState(false);
-  // Bumped by the retry button to re-run the connect effect.
-  const [attempt, setAttempt] = useState(0);
+  const [connecting, setConnecting] = useState(false);
 
-  // Connect on mount (no button): joining the room spins up the bot and keeps the
-  // session healthy, so the transcript starts (and keeps) flowing right away. The
-  // audio itself stays muted until the listener presses play.
-  useEffect(() => {
-    let cancelled = false;
-    const connect = async () => {
-      try {
-        // Ask the server to spin up (or reuse) the translator bot. For original
-        // audio this just ensures the default bridge (and its transcript) runs.
-        const tr = (await fetch("/api/livekit/translate", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ sessionId: docId, targetLanguage: translateTarget }),
-        }).then((r) => r.json())) as TranslateResp;
-        if (tr.error) throw new Error(tr.error);
+  const start = useCallback(async () => {
+    setError(null);
+    setConnecting(true);
+    try {
+      // Ask the server to spin up (or reuse) the translator bot. For original
+      // audio this just ensures the default bridge (and its transcript) is running.
+      const tr = (await fetch("/api/livekit/translate", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ sessionId: docId, targetLanguage: translateTarget }),
+      }).then((r) => r.json())) as TranslateResp;
+      if (tr.error) throw new Error(tr.error);
 
-        const identity = `attendee-${Math.random().toString(36).slice(2, 8)}`;
-        const tk = (await fetch("/api/livekit/token", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ room: docId, identity, role: "attendee" }),
-        }).then((r) => r.json())) as TokenResp;
-        if (tk.error) throw new Error(tk.error);
-        if (!tk.token || !tk.serverUrl) {
-          throw new Error("Incomplete response from server");
-        }
-        if (!isOriginal && !tr.translatorIdentity) {
-          throw new Error("Incomplete response from server");
-        }
-
-        if (cancelled) return;
-        setConn({
-          token: tk.token,
-          serverUrl: tk.serverUrl,
-          translatorIdentity: isOriginal ? null : tr.translatorIdentity ?? null,
-        });
-      } catch (e) {
-        if (!cancelled) setError(e instanceof Error ? e.message : String(e));
+      const identity = `attendee-${Math.random().toString(36).slice(2, 8)}`;
+      const tk = (await fetch("/api/livekit/token", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ room: docId, identity, role: "attendee" }),
+      }).then((r) => r.json())) as TokenResp;
+      if (tk.error) throw new Error(tk.error);
+      if (!tk.token || !tk.serverUrl) {
+        throw new Error("Incomplete response from server");
       }
-    };
-    void connect();
-    return () => {
-      cancelled = true;
-    };
-  }, [docId, translateTarget, isOriginal, attempt]);
+      if (!isOriginal && !tr.translatorIdentity) {
+        throw new Error("Incomplete response from server");
+      }
 
-  // The control bar reflects the connection: error (with retry), the live audio
-  // controls once connected, or a connecting hint. The transcript is always shown
-  // below it, since it reads Yjs and needs no LiveKit connection.
-  let controls: React.ReactElement;
+      setConn({
+        token: tk.token,
+        serverUrl: tk.serverUrl,
+        translatorIdentity: isOriginal ? null : tr.translatorIdentity ?? null,
+      });
+    } catch (e) {
+      setError(e instanceof Error ? e.message : String(e));
+    } finally {
+      setConnecting(false);
+    }
+  }, [docId, translateTarget, isOriginal]);
+
   if (error) {
-    controls = (
-      <div className="flex flex-col gap-1 items-start text-sm">
+    return (
+      <div className="flex flex-col gap-2 items-start text-sm">
         <p className="text-red-600 dark:text-red-400">{s.liveAudioError}</p>
         <p className="text-xs text-gray-500">{error}</p>
         <button
           type="button"
           className="px-3 py-1 rounded bg-blue-500 text-white text-xs hover:bg-blue-600"
-          onClick={() => {
-            setError(null);
-            setConn(null);
-            setAttempt((a) => a + 1);
-          }}
+          onClick={() => void start()}
         >
           {s.retry}
         </button>
       </div>
     );
-  } else if (conn) {
-    controls = (
-      <LiveKitRoom
-        video={false}
-        audio={false}
-        token={conn.token}
-        serverUrl={conn.serverUrl}
-        connectOptions={{ autoSubscribe: false }}
-        onError={(e) => setError(e.message)}
-      >
-        <ListenAudio
-          docId={docId}
-          releaseTarget={translateTarget}
-          translatorIdentity={conn.translatorIdentity}
-          isOriginal={isOriginal}
-          audioOn={audioOn}
-          onToggleAudio={() => setAudioOn((v) => !v)}
-        />
-      </LiveKitRoom>
-    );
-  } else {
-    controls = (
-      <div className="flex items-center gap-2 text-xs text-gray-500">{s.connecting}</div>
+  }
+
+  if (!conn) {
+    return (
+      <div className="flex-1 flex flex-col items-center justify-center gap-3">
+        <button
+          type="button"
+          disabled={connecting}
+          className="px-4 py-2 rounded-full bg-blue-500 text-white hover:bg-blue-600 disabled:opacity-60 shadow"
+          onClick={() => void start()}
+        >
+          {connecting ? s.connecting : `🔊 ${s.listenLive}`}
+        </button>
+      </div>
     );
   }
 
   return (
-    <div className="flex flex-col gap-2 flex-1 min-h-0">
-      {controls}
-      <LiveTranscript langCode={langCode} />
-    </div>
+    <LiveKitRoom
+      video={false}
+      audio={false}
+      token={conn.token}
+      serverUrl={conn.serverUrl}
+      connectOptions={{ autoSubscribe: false }}
+      onError={(e) => setError(e.message)}
+      className="flex flex-col gap-2 flex-1 min-h-0"
+    >
+      <ListenInner
+        docId={docId}
+        langCode={langCode}
+        releaseTarget={translateTarget}
+        translatorIdentity={conn.translatorIdentity}
+        isOriginal={isOriginal}
+      />
+    </LiveKitRoom>
   );
 }
